chore(repo): ignore .codex/ and unblock the dead-code audit
`.codex/` was untracked only because all four files in it end in `.log`,
matched by the existing `*.log` rule. The first non-log file written there
would have become a commit candidate, which is why the remediation plan's
execution contract carried a manual "do not commit .codex files" reminder.
That reminder is deleted with the rule that replaces it.

`npm run audit:dead-code` could not run at all without a prior perf build: it
exited 2 loading tests/perf/playwright.config.ts, whose module body threw when
dist artifacts were missing. Knip's playwright plugin globs
playwright.config.{js,ts,mjs} at any depth and LOADS every match during
plugin-config discovery, before its own ignore filtering applies -- so the
knip.json ignore entry the plan proposed as an alternative is a measured no-op,
as is ignoring all of tests/perf/** or negating the file out of the plugin's
own config list. All three reproduce the identical exit 2.

The build assertion moves to tests/perf/assert-perf-builds.mjs, invoked as a
preflight in the first webServer command. Not globalSetup: Playwright starts
every webServer before globalSetup runs, so a missing build would have
surfaced as a 120s vite-preview health-check timeout instead of the one-line
"run npm run perf:gate" message. The provenance print and staleness warning
move with it -- they called statSync on the same artifacts at module load, so
removing only the throw would have swapped a clear error for a bare ENOENT and
left the audit just as broken.

audit:dead-code now completes. It exits 1 on a pre-existing backlog (21 unused
files, 20 unlisted dependencies, ...) which is out of scope here, and it is
still wired to no CI job or hook -- this makes the command runnable, not a gate.
